requirements.txt
Flask==2.0.1
Werkzeug==2.2.2
Requests==2.31.0
anthropic>=0.25.0
openai>=1.30.0
google-generativeai==0.3.2
